+++ Assets/Scenes/MaizeBlue.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: MaizeBlue
   m_EditorClassIdentifier: 
   colorMode: 3
-  topLeft: {r: 0.92156863, g: 0.9490196, b: 0.98039216, a: 1}
-  topRight: {r: 0.92156863, g: 0.9490196, b: 0.98039216, a: 1}
+  topLeft: {r: 0.98039216, g: 0.78039217, b: 0.28235295, a: 1}
+  topRight: {r: 0.98039216, g: 0.78039217, b: 0.28235295, a: 1}
   bottomLeft: {r: 0.5137255, g: 0.5647059, b: 0.98039216, a: 1}
   bottomRight: {r: 0.5137255, g: 0.5647059, b: 0.98039216, a: 1}
