+++ Assets/Scripts/PlayerView/MainPausedMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7d9d4e101da406a4ea523aae46543b2b
+guid: fdcf0f0137842f84ba4248be20e6c4ac
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
