+++ Assets/Scripts/PlayerView/PlayerMenu.cs
@@ -1,0 +1,40 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerMenu : MonoBehaviour
+{
+    public GameObject playerMenu;
+    public GameObject player;
+    public GameObject minimap;
+    public bool ispaused;
+
+    public void DeactivateMenu()
+    {
+        ispaused = !ispaused;
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.F2) || Input.GetKeyDown(KeyCode.F3))
+        {
+            if (ispaused)
+            {
+                Debug.Log("Hello");
+                playerMenu.SetActive(false);
+                player.SetActive(true);
+                minimap.SetActive(true);
+                ispaused = !ispaused;
+            }
+            else
+            {
+                Debug.Log("Hello"); 
+                player.SetActive(false);
+                minimap.SetActive(false);
+                playerMenu.SetActive(true);
+                ispaused = !ispaused;
+            }
+        }
+    }
+}
