+++ Assets/Scripts/PlayerView/PlayerMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7a13c038b73c1f04998e6970f9233bac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
