+++ ProjectSettings/EditorBuildSettings.asset
@@ -11,7 +11,4 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Main.unity
     guid: b29595dc884434cf3bfe1dc95780d0a2
-  - enabled: 0
-    path: 
-    guid: 00000000000000000000000000000000
   m_configObjects: {}
