+++ UserSettings/EditorUserSettings.asset
@@ -6,7 +6,6 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
-
       value: 224247031146466804031c2e150d5e0403051d1a2d2b2d7f1e2a183de7f37b0bebea26e5e2143e282c10ea0c053d0a00d60f0404b81e04020517
       flags: 0
     RecentlyUsedScenePath-1:
@@ -34,7 +33,6 @@ EditorUserSettings:
       value: 2242470311464672001e033004265b5037040b2f383b690324240d3fe7ce352cf7f533d9e63934730a01ea32012d4e0cfb071907f3250b1f041cf83e10fc1b2c35c015ca54dd1115df00
       flags: 0
     RecentlyUsedScenePath-9:
-
       value: 22424703114646680e0b0227036c72111f19563f22213229
       flags: 0
     vcSharedLogLevel:
